+++ src/entities/comment/model/mutations.ts
@@ -1,6 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { createComment, deleteComment, updateComment, likeComment } from '../api';
 import type { Comment } from '../types';
+import { MOCK_MODE } from '../../../shared/config';
 
 export function useAddCommentMutation(postId: number | null) {
   const qc = useQueryClient();
@@ -26,13 +27,22 @@ export function useAddCommentMutation(postId: number | null) {
         ],
       });
 
-      return { key, prev };
+      return { key, prev, tempId };
     },
     onError: (_e, _v, ctx) => {
       if (ctx) qc.setQueryData(ctx.key, ctx.prev);
     },
+    onSuccess: (serverComment, _vars, ctx) => {
+      if (!ctx) return;
+      const { key, prev, tempId } = ctx as any;
+      const cur = qc.getQueryData<any>(key);
+      if (!cur?.comments) return;
+      qc.setQueryData(key, {
+        comments: cur.comments.map((c: Comment) => (c.id === tempId ? serverComment : c)),
+      });
+    },
     onSettled: () => {
-      if (postId) qc.invalidateQueries({ queryKey: ['comments', postId] });
+      if (!MOCK_MODE && postId) qc.invalidateQueries({ queryKey: ['comments', postId] });
     },
   });
 }
@@ -59,7 +69,7 @@ export function useUpdateCommentMutation(postId: number | null) {
       if (ctx) qc.setQueryData(ctx.key, ctx.prev);
     },
     onSettled: () => {
-      if (postId) qc.invalidateQueries({ queryKey: ['comments', postId] });
+      if (!MOCK_MODE && postId) qc.invalidateQueries({ queryKey: ['comments', postId] });
     },
   });
 }
@@ -84,16 +94,17 @@ export function useDeleteCommentMutation(postId: number | null) {
       if (ctx) qc.setQueryData(ctx.key, ctx.prev);
     },
     onSettled: () => {
-      if (postId) qc.invalidateQueries({ queryKey: ['comments', postId] });
+      if (!MOCK_MODE && postId) qc.invalidateQueries({ queryKey: ['comments', postId] });
     },
   });
 }
 
+// 서버가 likes를 저장하지 않을 수 있어 '낙관적 업데이트'만 유지
 export function useLikeCommentMutation(postId: number | null) {
   const qc = useQueryClient();
   return useMutation({
     mutationFn: (input: { id: number; nextLikes: number }) =>
-      likeComment({ id: input.id, likes: input.nextLikes }),
+      likeComment({ id: input.id, likes: input.nextLikes }), // 서버가 무시해도 OK
     onMutate: async ({ id }) => {
       if (!postId) return;
       await qc.cancelQueries({ queryKey: ['comments', postId] });
@@ -112,7 +123,8 @@ export function useLikeCommentMutation(postId: number | null) {
       if (ctx) qc.setQueryData(ctx.key, ctx.prev);
     },
     onSettled: () => {
-      if (postId) qc.invalidateQueries({ queryKey: ['comments', postId] });
+      // MOCK_MODE면 invalidate 안 함 (원복 방지)
+      if (!MOCK_MODE && postId) qc.invalidateQueries({ queryKey: ['comments', postId] });
     },
   });
 }
