+++ src/shared/config.ts
@@ -1,0 +1,1 @@
+export const MOCK_MODE = true; // DummyJSON 같은 비지속 API 사용할 때 true
